refactor: 💡 Modularize cli
